+++ src/modules/contacts/dtos/create-contact-method-email.dto.ts
@@ -1,8 +1,8 @@
 import { IsString } from 'class-validator';
+import { ContactEmailDto } from './contact-method-email.dto';
 import { ApiProperty } from '@nestjs/swagger';
-import { ContactPhoneDto } from './contact-phone.dto';
 
-export class CreateContactPhoneDto extends ContactPhoneDto {
+export class CreateContactEmailDto extends ContactEmailDto {
   @ApiProperty()
   @IsString()
   contactId: string;
