+++ src/modules/contacts/mappers/contact-email.mapper.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { ContactEmailDto } from '../dtos/contact-email.dto';
+import { ContactEmailDto } from '../dtos/contact-method-email.dto';
 import { ContactEmail } from '../entities/contact-emails.entity';
 
 @Injectable()
